Crate par_iter

Fork of rayon to use parallel iterators with chili. See par-core for more details.

Rayon is a data-parallelism library that makes it easy to convert sequential computations into parallel.

It is lightweight and convenient for introducing parallelism into existing code. It guarantees data-race free executions and takes advantage of parallelism when sensible, based on work-load at runtime.

How to use Rayon

There are two ways to use Rayon:

High-level parallel constructs are the simplest way to use Rayon and also typically the most efficient.

Parallel iterators make it easy to convert a sequential iterator to execute in parallel.

The ParallelIterator trait defines general methods for all parallel iterators.

The IndexedParallelIterator trait adds methods for iterators that support random access.

The par_sort method sorts &mut [T] slices (or vectors) in parallel.

par_extend can be used to efficiently grow collections with items produced by a parallel iterator.

Custom tasks let you divide your work into parallel tasks yourself.

join is used to subdivide a task into two pieces.

scope creates a scope within which you can create any number of parallel tasks.

ThreadPoolBuilder can be used to create your own thread pools or customize the global one.

Basic usage and the Rayon prelude

First, you will need to add rayon to your Cargo.toml.

Next, to use parallel iterators or the other high-level methods, you need to import several traits. Those traits are bundled into the module par_iter::prelude. It is recommended that you import all of these traits at once by adding use par_iter::prelude::* at the top of each module that uses Rayon methods.

These traits give you access to the par_iter method which provides parallel implementations of many iterative functions such as map, for_each, filter, fold, and more.

Crate Layout

Rayon extends many of the types found in the standard library with parallel iterator implementations. The modules in the rayon crate mirror std itself: so, e.g., the option module in Rayon contains parallel iterators for the Option type, which is found in the option module of std. Similarly, the collections module in Rayon offers parallel iterator types for the collections from std. You will rarely need to access these submodules unless you need to name iterator types explicitly.

Targets without threading

Rayon has limited support for targets without std threading implementations. See the [rayon_core] documentation for more information about its global fallback.

Other questions?

See the Rayon FAQ.

Modules

array

Parallel iterator types for arrays ([T; N])

collections

Parallel iterator types for standard collections

iter

Traits for writing parallel programs using an iterator-style interface

option

Parallel iterator types for options

prelude

The rayon prelude imports the various ParallelIterator traits. The intention is that one can include use par_iter::prelude::* and have easy access to the various traits and methods you will need.

range

Parallel iterator types for ranges, the type for values created by a..b expressions

range_inclusive

Parallel iterator types for inclusive ranges, the type for values created by a..=b expressions

result

Parallel iterator types for results

slice

Parallel iterator types for slices

str

Parallel iterator types for strings

string

This module contains the parallel iterator types for owned strings (String). You will rarely need to interact with it directly unless you have need to name one of the iterator types.

vec

Parallel iterator types for vectors (Vec<T>)